인선: HealthyDTO 수정

+++ src/bean/healthy/HealthyDTO.java
@@ -12,73 +12,61 @@ create table t_healthy_board (
 	);
 */
 public class HealthyDTO {
-	private long num;
+	private int num;
 	private String title;
 	private String content;
 	private String writer;
 	private Timestamp regdate;
 	private Timestamp updateDate;
 	
-	
-	public long getNum() {
+
+	public int getNum() {
 		return num;
 	}
-
 
 	public String getTitle() {
 		return title;
 	}
 
-
 	public String getContent() {
 		return content;
 	}
-
 
 	public String getWriter() {
 		return writer;
 	}
 
-
 	public Timestamp getRegdate() {
 		return regdate;
 	}
-
 
 	public Timestamp getUpdateDate() {
 		return updateDate;
 	}
 
-
-	public void setNum(long num) {
+	public void setNum(int num) {
 		this.num = num;
 	}
-
 
 	public void setTitle(String title) {
 		this.title = title;
 	}
 
-
 	public void setContent(String content) {
 		this.content = content;
 	}
-
 
 	public void setWriter(String writer) {
 		this.writer = writer;
 	}
 
-
 	public void setRegdate(Timestamp regdate) {
 		this.regdate = regdate;
 	}
 
-
 	public void setUpdateDate(Timestamp updateDate) {
 		this.updateDate = updateDate;
 	}
-
 
 	@Override
 	public String toString() {
@@ -86,5 +74,5 @@ public class HealthyDTO {
 				+ content + ", writer=" + writer + ", regdate=" + regdate + ", updateDate=" + updateDate + "]";
 	}
 	
-	
+
 }
